Minor changes to pathnames for consistency
