+++ Package.swift
@@ -4,13 +4,13 @@
 import PackageDescription
 
 let package = Package(
-    name: "IgisShellD",
+    name: "IgisShell",
     dependencies: [
     ],
     targets: [
         // Targets are the basic building blocks of a package. A target can define a module or a test suite.
         // Targets can depend on other targets in this package, and on products in packages which this package depends on.
         .target(
-            name: "IgisShellD"),
+            name: "IgisShell"),
     ]
 )
